test numpy funcs on patch
