+++ dascore/proc/aggregate.py
@@ -3,34 +3,13 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Sequence
-from functools import partial
 
 import numpy as np
 
-import dascore as dc
-from dascore.constants import PatchType
-from dascore.exceptions import ParameterError
+from dascore.constants import _AGG_FUNCS, PatchType
+from dascore.utils.array import _apply_aggregator
 from dascore.utils.docs import compose_docstring
-from dascore.utils.misc import iterate
-from dascore.utils.patch import get_dim_axis_value, patch_function
-from dascore.utils.time import (
-    dtype_time_like,
-    is_datetime64,
-    to_datetime64,
-    to_timedelta64,
-)
-
-_AGG_FUNCS = {
-    "mean": np.nanmean,
-    "median": np.nanmedian,
-    "min": np.nanmin,
-    "max": np.nanmax,
-    "sum": np.nansum,
-    "std": np.nanstd,
-    "first": partial(np.take, indices=0),
-    "last": partial(np.take, indices=-1),
-}
-
+from dascore.utils.patch import patch_function
 
 AGG_DOC_STR = """
 patch
@@ -55,37 +34,6 @@ Notes
 See [`Patch.aggregate`](`dascore.Patch.aggregate`) for examples
 and more details.
 """
-
-
-def _get_new_coord(coord, dim_reduce):
-    """Get the new coordinate."""
-
-    def _maybe_handle_datatypes(func, data):
-        """Maybe handle the complexity of date times here."""
-        try:  # First try function directly
-            out = func(data)
-        except Exception:  # Fall back to floats and re-packing.
-            float_data = dc.to_float(data)
-            dfunc = to_datetime64 if is_datetime64(data) else to_timedelta64
-            out = dfunc(func(float_data))
-        return np.atleast_1d(out)
-
-    if dim_reduce == "empty":
-        new_coord = coord.update(shape=(1,), start=None, stop=None, data=None)
-    elif dim_reduce == "squeeze":
-        return None
-    elif (func := _AGG_FUNCS.get(dim_reduce)) or callable(dim_reduce):
-        func = dim_reduce if callable(dim_reduce) else func
-        coord_data = coord.data
-        if dtype_time_like(coord_data):
-            result = _maybe_handle_datatypes(func, coord_data)
-        else:
-            result = func(coord.data)
-        new_coord = coord.update(data=result)
-    else:
-        msg = "dim_reduce must be 'empty', 'squeeze' or valid aggregator."
-        raise ParameterError(msg)
-    return new_coord
 
 
 @patch_function()
@@ -135,20 +83,7 @@ def aggregate(
     ... )
     """
     func = _AGG_FUNCS.get(method, method)
-    data = patch.data
-    dims = tuple(iterate(patch.dims if dim is None else dim))
-    dfo = get_dim_axis_value(patch, args=dims, allow_multiple=True)
-    # Iter all specified dimensions.
-    for dim, axis, value in dfo:
-        new_coord = _get_new_coord(patch.get_coord(dim), dim_reduce=dim_reduce)
-        if new_coord is None:
-            coords = patch.coords.drop_coords(dim)[0]
-            data = func(data, axis=axis)
-        else:
-            coords = patch.coords.update(**{dim: new_coord})
-            data = np.expand_dims(func(data, axis=axis), axis)
-        patch = patch.new(data=data, coords=coords)
-    return patch
+    return _apply_aggregator(patch, dim, func, dim_reduce)
 
 
 @patch_function()
